exchange_rate_service/pyproject.toml
[tool.poetry]
name = "exchange-rate-service"
version = "0.1.0"
description = ""
authors = ["AlexSitohov <lilinfinitemob@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
redis = "^5.0.1"
yoyo-migrations = "^8.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pylint = "^3.0.3"
black = "^24.1.1"
yoyo-migrations = "^8.2.0"
pytest = "^8.0.0"
testcontainers = "^3.7.1"
httpx = "^0.26.0"
pytest-asyncio = "^0.23.4"
coverage = "^7.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
